Eliminated compilation warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@19680 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/terrain_brute/object/brute.cpp
@@ -1701,7 +1701,8 @@ void csTerrainObject::CastShadows (iMovable* movable, iFrustumView* fview)
     {
       if (i % 10000 == 0)
       {
-        printf ("%d out of %d\n", i, staticLights.Length ());
+	printf ("%lu out of %lu\n", (unsigned long)i,
+	  (unsigned long)staticLights.Length ());
 	fflush (stdout);
       }
     }
